refactor: split tui and daemon runtime concerns
