+++ internal/daemon/server_poll.go
@@ -1,0 +1,137 @@
+package daemon
+
+import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+
+	"github.com/janekbaraniewski/openusage/internal/core"
+)
+
+func (s *Service) runPollLoop(ctx context.Context) {
+	ticker := time.NewTicker(s.cfg.PollInterval)
+	defer ticker.Stop()
+
+	s.infof("poll_loop_start", "interval=%s", s.cfg.PollInterval)
+	s.pollProviders(ctx)
+	for {
+		select {
+		case <-ctx.Done():
+			s.infof("poll_loop_stop", "reason=context_done")
+			return
+		case <-ticker.C:
+			s.pollProviders(ctx)
+		}
+	}
+}
+
+func (s *Service) pollProviders(ctx context.Context) {
+	if s == nil || s.quotaIngest == nil {
+		return
+	}
+	started := time.Now()
+
+	accounts, modelNorm, err := LoadAccountsAndNorm()
+	if err != nil {
+		if s.shouldLog("poll_config_warning", 20*time.Second) {
+			s.warnf("poll_config_warning", "error=%v", err)
+		}
+		return
+	}
+	if len(accounts) == 0 {
+		if s.shouldLog("poll_no_accounts", 30*time.Second) {
+			s.infof("poll_skipped", "reason=no_enabled_accounts")
+		}
+		return
+	}
+
+	type providerResult struct {
+		accountID string
+		snapshot  core.UsageSnapshot
+	}
+
+	results := make(chan providerResult, len(accounts))
+	var wg sync.WaitGroup
+
+	for _, acct := range accounts {
+		wg.Add(1)
+		go func(account core.AccountConfig) {
+			defer wg.Done()
+
+			provider, ok := s.providerByID[account.Provider]
+			if !ok {
+				results <- providerResult{
+					accountID: account.ID,
+					snapshot: core.UsageSnapshot{
+						ProviderID: account.Provider,
+						AccountID:  account.ID,
+						Timestamp:  time.Now().UTC(),
+						Status:     core.StatusError,
+						Message:    fmt.Sprintf("no provider adapter registered for %q (restart/reinstall telemetry daemon if recently added)", account.Provider),
+					},
+				}
+				return
+			}
+
+			fetchCtx, cancel := context.WithTimeout(ctx, 8*time.Second)
+			defer cancel()
+
+			snap, fetchErr := provider.Fetch(fetchCtx, account)
+			if fetchErr != nil {
+				snap = core.UsageSnapshot{
+					ProviderID: account.Provider,
+					AccountID:  account.ID,
+					Timestamp:  time.Now().UTC(),
+					Status:     core.StatusError,
+					Message:    fetchErr.Error(),
+				}
+			}
+			snap = core.NormalizeUsageSnapshotWithConfig(snap, modelNorm)
+			results <- providerResult{accountID: account.ID, snapshot: snap}
+		}(acct)
+	}
+
+	go func() {
+		wg.Wait()
+		close(results)
+	}()
+
+	snapshots := make(map[string]core.UsageSnapshot, len(accounts))
+	statusCounts := map[core.Status]int{}
+	errorCount := 0
+	for result := range results {
+		snapshots[result.accountID] = result.snapshot
+		statusCounts[result.snapshot.Status]++
+		if result.snapshot.Status == core.StatusError {
+			errorCount++
+		}
+	}
+	if len(snapshots) == 0 {
+		return
+	}
+
+	ingestCtx, cancel := context.WithTimeout(ctx, 12*time.Second)
+	defer cancel()
+	ingestErr := s.ingestQuotaSnapshots(ingestCtx, snapshots)
+	if ingestErr != nil && s.shouldLog("poll_ingest_warning", 10*time.Second) {
+		s.warnf("poll_ingest_warning", "error=%v", ingestErr)
+	}
+
+	durationMs := time.Since(started).Milliseconds()
+	if ingestErr != nil || errorCount > 0 || s.shouldLog("poll_cycle_info", 45*time.Second) {
+		s.infof(
+			"poll_cycle",
+			"duration_ms=%d accounts=%d snapshots=%d status_ok=%d status_auth=%d status_limited=%d status_error=%d status_unknown=%d ingest_error=%t",
+			durationMs,
+			len(accounts),
+			len(snapshots),
+			statusCounts[core.StatusOK],
+			statusCounts[core.StatusAuth],
+			statusCounts[core.StatusLimited],
+			statusCounts[core.StatusError],
+			statusCounts[core.StatusUnknown],
+			ingestErr != nil,
+		)
+	}
+}
