+++ internal/tui/model.go
@@ -2,7 +2,6 @@ package tui
 
 import (
 	"fmt"
-	"log"
 	"sort"
 	"strings"
 	"time"
@@ -266,901 +265,7 @@ type integrationInstallResultMsg struct {
 	Err           error
 }
 
-func (m Model) persistThemeCmd(themeName string) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return themePersistedMsg{err: fmt.Errorf("theme service unavailable")}
-		}
-		err := m.services.SaveTheme(themeName)
-		if err != nil {
-			log.Printf("theme persist: %v", err)
-		}
-		return themePersistedMsg{err: err}
-	}
-}
-
-func (m Model) persistDashboardPrefsCmd() tea.Cmd {
-	providers := m.dashboardConfigProviders()
-	return func() tea.Msg {
-		if m.services == nil {
-			return dashboardPrefsPersistedMsg{err: fmt.Errorf("dashboard settings service unavailable")}
-		}
-		err := m.services.SaveDashboardProviders(providers)
-		if err != nil {
-			log.Printf("dashboard settings persist: %v", err)
-		}
-		return dashboardPrefsPersistedMsg{err: err}
-	}
-}
-
-func (m Model) persistDashboardViewCmd() tea.Cmd {
-	view := string(m.configuredDashboardView())
-	return func() tea.Msg {
-		if m.services == nil {
-			return dashboardViewPersistedMsg{err: fmt.Errorf("dashboard view service unavailable")}
-		}
-		err := m.services.SaveDashboardView(view)
-		if err != nil {
-			log.Printf("dashboard view persist: %v", err)
-		}
-		return dashboardViewPersistedMsg{err: err}
-	}
-}
-
-func (m Model) persistDashboardWidgetSectionsCmd() tea.Cmd {
-	sections := m.dashboardWidgetSectionConfigEntries()
-	return func() tea.Msg {
-		if m.services == nil {
-			return dashboardWidgetSectionsPersistedMsg{err: fmt.Errorf("dashboard sections service unavailable")}
-		}
-		err := m.services.SaveDashboardWidgetSections(sections)
-		if err != nil {
-			log.Printf("dashboard widget sections persist: %v", err)
-		}
-		return dashboardWidgetSectionsPersistedMsg{err: err}
-	}
-}
-
-func (m Model) persistDashboardHideSectionsWithNoDataCmd() tea.Cmd {
-	hide := m.hideSectionsWithNoData
-	return func() tea.Msg {
-		if m.services == nil {
-			return dashboardHideSectionsWithNoDataPersistedMsg{err: fmt.Errorf("dashboard empty-state service unavailable")}
-		}
-		err := m.services.SaveDashboardHideSectionsWithNoData(hide)
-		if err != nil {
-			log.Printf("dashboard hide sections with no data persist: %v", err)
-		}
-		return dashboardHideSectionsWithNoDataPersistedMsg{err: err}
-	}
-}
-
-func (m Model) persistTimeWindowCmd(window string) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return timeWindowPersistedMsg{err: fmt.Errorf("time window service unavailable")}
-		}
-		err := m.services.SaveTimeWindow(window)
-		if err != nil {
-			log.Printf("time window persist: %v", err)
-		}
-		return timeWindowPersistedMsg{err: err}
-	}
-}
-
-func (m Model) validateKeyCmd(accountID, providerID, apiKey string) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return validateKeyResultMsg{AccountID: accountID, Valid: false, Error: "validation service unavailable"}
-		}
-		valid, errMsg := m.services.ValidateAPIKey(accountID, providerID, apiKey)
-		return validateKeyResultMsg{AccountID: accountID, Valid: valid, Error: errMsg}
-	}
-}
-
-func (m Model) saveCredentialCmd(accountID, apiKey string) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return credentialSavedMsg{AccountID: accountID, Err: fmt.Errorf("credential service unavailable")}
-		}
-		err := m.services.SaveCredential(accountID, apiKey)
-		return credentialSavedMsg{AccountID: accountID, Err: err}
-	}
-}
-
-func (m Model) deleteCredentialCmd(accountID string) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return credentialDeletedMsg{AccountID: accountID, Err: fmt.Errorf("credential service unavailable")}
-		}
-		err := m.services.DeleteCredential(accountID)
-		return credentialDeletedMsg{AccountID: accountID, Err: err}
-	}
-}
-
-func (m Model) installIntegrationCmd(id integrations.ID) tea.Cmd {
-	return func() tea.Msg {
-		if m.services == nil {
-			return integrationInstallResultMsg{IntegrationID: id, Err: fmt.Errorf("integration service unavailable")}
-		}
-		statuses, err := m.services.InstallIntegration(id)
-		return integrationInstallResultMsg{
-			IntegrationID: id,
-			Statuses:      statuses,
-			Err:           err,
-		}
-	}
-}
-
 func (m Model) Init() tea.Cmd { return tickCmd() }
-
-func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tickMsg:
-		m.animFrame++
-		return m, tickCmd()
-
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-		return m, nil
-
-	case DaemonStatusMsg:
-		m.daemon.status = msg.Status
-		m.daemon.message = msg.Message
-		if msg.Status == DaemonRunning {
-			m.daemon.installing = false
-		}
-		return m, nil
-
-	case AppUpdateMsg:
-		m.daemon.appUpdateCurrent = strings.TrimSpace(msg.CurrentVersion)
-		m.daemon.appUpdateLatest = strings.TrimSpace(msg.LatestVersion)
-		m.daemon.appUpdateHint = strings.TrimSpace(msg.UpgradeHint)
-		return m, nil
-
-	case daemonInstallResultMsg:
-		m.daemon.installing = false
-		if msg.err != nil {
-			m.daemon.status = DaemonError
-			m.daemon.message = msg.err.Error()
-		} else {
-			m.daemon.installDone = true
-			m.daemon.status = DaemonStarting
-		}
-		return m, nil
-
-	case SnapshotsMsg:
-		msgWindow := msg.TimeWindow
-		if msgWindow == "" {
-			msgWindow = core.TimeWindow30d
-		}
-		if msgWindow != m.timeWindow {
-			return m, nil
-		}
-		if msg.RequestID > 0 && msg.RequestID < m.lastSnapshotRequestID {
-			return m, nil
-		}
-		if m.refreshing && m.hasData && !snapshotsReady(msg.Snapshots) {
-			return m, nil
-		}
-		m.snapshots = msg.Snapshots
-		m.refreshing = false
-		if msg.RequestID > m.lastSnapshotRequestID {
-			m.lastSnapshotRequestID = msg.RequestID
-		}
-		if len(msg.Snapshots) > 0 || snapshotsReady(msg.Snapshots) {
-			m.hasData = true
-			m.daemon.status = DaemonRunning
-		}
-		for id, snap := range m.snapshots {
-			info := computeDisplayInfo(snap, dashboardWidget(snap.ProviderID))
-			if info.reason != "" {
-				snap.EnsureMaps()
-				snap.Diagnostics["display_branch"] = info.reason
-				m.snapshots[id] = snap
-			}
-		}
-		m.ensureSnapshotProvidersKnown()
-		m.rebuildSortedIDs()
-		return m, nil
-
-	case dashboardPrefsPersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "save failed"
-		} else {
-			m.settings.status = "saved"
-		}
-		return m, nil
-
-	case dashboardViewPersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "view save failed"
-		} else {
-			m.settings.status = "view saved"
-		}
-		return m, nil
-
-	case dashboardWidgetSectionsPersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "section save failed"
-		} else {
-			m.settings.status = "sections saved"
-		}
-		return m, nil
-
-	case dashboardHideSectionsWithNoDataPersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "empty-state save failed"
-		} else {
-			m.settings.status = "empty-state saved"
-		}
-		return m, nil
-
-	case themePersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "theme save failed"
-		} else {
-			m.settings.status = "theme saved"
-		}
-		return m, nil
-
-	case timeWindowPersistedMsg:
-		if msg.err != nil {
-			m.settings.status = "time window save failed"
-		} else {
-			m.settings.status = "time window saved"
-		}
-		return m, nil
-
-	case validateKeyResultMsg:
-		if msg.Valid {
-			m.settings.apiKeyStatus = "valid ✓ — saving..."
-			return m, m.saveCredentialCmd(msg.AccountID, m.settings.apiKeyInput)
-		}
-		m.settings.apiKeyStatus = "invalid ✗"
-		if msg.Error != "" {
-			errMsg := msg.Error
-			if len(errMsg) > 40 {
-				errMsg = errMsg[:37] + "..."
-			}
-			m.settings.apiKeyStatus = "invalid: " + errMsg
-		}
-		return m, nil
-
-	case credentialSavedMsg:
-		if msg.Err != nil {
-			m.settings.apiKeyStatus = "save failed"
-		} else {
-			m.settings.apiKeyStatus = "saved ✓"
-			apiKey := m.settings.apiKeyInput
-			m.settings.apiKeyEditing = false
-			m.settings.apiKeyInput = ""
-
-			// Register account with engine if callback is set
-			if m.onAddAccount != nil {
-				providerID := m.accountProviders[msg.AccountID]
-				acct := core.AccountConfig{
-					ID:       msg.AccountID,
-					Provider: providerID,
-					Auth:     "api_key",
-					Token:    apiKey,
-				}
-				m.onAddAccount(acct)
-			}
-
-			// Ensure the provider shows in the UI
-			if m.providerOrderIndex(msg.AccountID) < 0 {
-				m.providerOrder = append(m.providerOrder, msg.AccountID)
-				m.providerEnabled[msg.AccountID] = true
-			}
-			m.refreshing = true
-		}
-		return m, nil
-
-	case credentialDeletedMsg:
-		if msg.Err != nil {
-			m.settings.status = "delete failed"
-		} else {
-			m.settings.status = "key deleted"
-		}
-		return m, nil
-
-	case integrationInstallResultMsg:
-		m.settings.integrationStatus = msg.Statuses
-		if msg.Err != nil {
-			errMsg := msg.Err.Error()
-			if len(errMsg) > 80 {
-				errMsg = errMsg[:77] + "..."
-			}
-			m.settings.status = "integration install failed: " + errMsg
-		} else {
-			m.settings.status = "integration installed"
-		}
-		return m, nil
-
-	case tea.KeyMsg:
-		if !m.hasData {
-			return m.handleSplashKey(msg)
-		}
-		return m.handleKey(msg)
-	case tea.MouseMsg:
-		return m.handleMouse(msg)
-	}
-	return m, nil
-}
-
-func (m Model) handleSplashKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "q", "ctrl+c":
-		return m, tea.Quit
-	case "enter":
-		if (m.daemon.status == DaemonNotInstalled || m.daemon.status == DaemonOutdated) && !m.daemon.installing {
-			m.daemon.installing = true
-			m.daemon.message = "Setting up background helper..."
-			return m, m.installDaemonCmd()
-		}
-	}
-	return m, nil
-}
-
-func (m Model) handleMouse(msg tea.MouseMsg) (tea.Model, tea.Cmd) {
-	if m.settings.show {
-		return m.handleSettingsMouse(msg)
-	}
-	if m.showHelp {
-		return m, nil
-	}
-	if m.filter.active || m.analyticsFilter.active {
-		return m, nil
-	}
-	if msg.Action != tea.MouseActionPress {
-		return m, nil
-	}
-
-	// Handle left-click tile selection in grid/list mode.
-	if msg.Button == tea.MouseButtonLeft {
-		return m.handleMouseClick(msg)
-	}
-
-	scroll := 0
-	switch msg.Button {
-	case tea.MouseButtonWheelUp:
-		scroll = -m.mouseScrollStep()
-	case tea.MouseButtonWheelDown:
-		scroll = m.mouseScrollStep()
-	default:
-		return m, nil
-	}
-
-	if m.screen != screenDashboard {
-		return m, nil
-	}
-
-	if m.mode == modeDetail {
-		m.detailOffset += scroll
-		if m.detailOffset < 0 {
-			m.detailOffset = 0
-		}
-		return m, nil
-	}
-
-	if m.mode == modeList && m.shouldUseWidgetScroll() {
-		m.tileOffset += scroll
-		if m.tileOffset < 0 {
-			m.tileOffset = 0
-		}
-		return m, nil
-	}
-
-	if m.mode == modeList && m.shouldUsePanelScroll() {
-		m.tileOffset += scroll
-		if m.tileOffset < 0 {
-			m.tileOffset = 0
-		}
-		return m, nil
-	}
-
-	if m.mode == modeList && m.activeDashboardView() == dashboardViewSplit {
-		step := 1
-		if scroll < 0 {
-			step = -1
-		}
-		next := m.cursor + step
-		ids := m.filteredIDs()
-		if next < 0 {
-			next = 0
-		}
-		if next >= len(ids) {
-			next = len(ids) - 1
-		}
-		if next < 0 {
-			next = 0
-		}
-		m.cursor = next
-	}
-
-	return m, nil
-}
-
-func (m Model) handleSettingsMouse(msg tea.MouseMsg) (tea.Model, tea.Cmd) {
-	if msg.Action != tea.MouseActionPress {
-		return m, nil
-	}
-	if m.settings.tab != settingsTabWidgetSections {
-		return m, nil
-	}
-
-	scroll := 0
-	switch msg.Button {
-	case tea.MouseButtonWheelUp:
-		scroll = -m.mouseScrollStep()
-	case tea.MouseButtonWheelDown:
-		scroll = m.mouseScrollStep()
-	default:
-		return m, nil
-	}
-
-	m.settings.previewOffset += scroll
-	if m.settings.previewOffset < 0 {
-		m.settings.previewOffset = 0
-	}
-	return m, nil
-}
-
-// handleMouseClick selects the tile under the mouse cursor when clicked.
-func (m Model) handleMouseClick(msg tea.MouseMsg) (tea.Model, tea.Cmd) {
-	if m.screen != screenDashboard || m.mode != modeList {
-		return m, nil
-	}
-
-	ids := m.filteredIDs()
-	if len(ids) == 0 {
-		return m, nil
-	}
-
-	view := m.activeDashboardView()
-	switch view {
-	case dashboardViewGrid, dashboardViewStacked:
-	default:
-		return m, nil
-	}
-
-	contentH := m.height - 3 // header (2 lines) + footer (1 line)
-	if contentH < 5 {
-		contentH = 5
-	}
-	cols, tileW, tileMaxH := m.tileGrid(m.width, contentH, len(ids))
-	if view == dashboardViewStacked {
-		cols = 1
-	}
-
-	// Header is 2 lines (brand + separator), content starts at Y=2.
-	headerH := 2
-	clickY := msg.Y - headerH
-	clickX := msg.X - 1 // 1-char left padding in content
-
-	if clickX < 0 || clickY < 0 {
-		return m, nil
-	}
-
-	// Determine column from X position.
-	cellW := tileW + tileBorderH + tileGapH
-	if cellW <= 0 {
-		return m, nil
-	}
-	col := clickX / cellW
-	if col >= cols {
-		return m, nil
-	}
-
-	// Determine row from Y position.
-	// We need to figure out the visible row height and account for scroll.
-	// Each tile row is tileMaxH + tileBorderV lines, rows separated by tileGapV.
-	// For single-column (stacked), tileMaxH=0 which means variable height; use an
-	// approximate row height in that case.
-	var rowH int
-	if tileMaxH > 0 {
-		rowH = tileMaxH + tileBorderV
-	} else {
-		// Variable height tiles — estimate from content area.
-		rowH = contentH
-		if len(ids) > 1 {
-			rowH = contentH / len(ids)
-		}
-		if rowH < tileMinHeight+tileBorderV {
-			rowH = tileMinHeight + tileBorderV
-		}
-	}
-
-	rowCell := rowH + tileGapV
-	if rowCell <= 0 {
-		return m, nil
-	}
-
-	// Account for scroll offset.
-	// In grid mode the view scrolls to keep the cursor row visible.
-	// We need to know which line the viewport starts at.
-	cursorRow := m.cursor / cols
-	totalRows := (len(ids) + cols - 1) / cols
-
-	// Build row offsets like renderTilesWithColumns does.
-	rowOffsets := make([]int, totalRows)
-	acc := 0
-	for r := 0; r < totalRows; r++ {
-		rowOffsets[r] = acc
-		acc += rowH
-		if r < totalRows-1 {
-			acc += tileGapV
-		}
-	}
-	totalLines := acc
-
-	rowScrollOffset := 0
-	if cols == 1 {
-		rowScrollOffset = m.tileOffset
-	}
-	scrollLine := 0
-	if cursorRow >= 0 && cursorRow < totalRows {
-		scrollLine = rowOffsets[cursorRow] + rowScrollOffset
-	}
-	if scrollLine > totalLines-contentH {
-		scrollLine = totalLines - contentH
-	}
-	if scrollLine < 0 {
-		scrollLine = 0
-	}
-
-	absY := clickY + scrollLine
-	row := -1
-	for r := 0; r < totalRows; r++ {
-		if absY >= rowOffsets[r] && absY < rowOffsets[r]+rowH {
-			row = r
-			break
-		}
-	}
-	if row < 0 {
-		return m, nil
-	}
-
-	idx := row*cols + col
-	if idx < 0 || idx >= len(ids) {
-		return m, nil
-	}
-
-	m.cursor = idx
-	m.tileOffset = 0
-	return m, nil
-}
-
-func (m Model) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	if msg.String() == "?" && !m.filter.active && !m.analyticsFilter.active && !m.settings.show {
-		m.showHelp = !m.showHelp
-		return m, nil
-	}
-	if m.showHelp {
-		m.showHelp = false
-		return m, nil
-	}
-
-	if m.settings.show {
-		return m.handleSettingsModalKey(msg)
-	}
-
-	if !m.filter.active && !m.analyticsFilter.active {
-		switch msg.String() {
-		case ",", "S":
-			m.openSettingsModal()
-			return m, nil
-		case "tab":
-			m.screen = m.nextScreen(1)
-			m.mode = modeList
-			m.detailOffset = 0
-			m.tileOffset = 0
-			return m, nil
-		case "shift+tab":
-			m.screen = m.nextScreen(-1)
-			m.mode = modeList
-			m.detailOffset = 0
-			m.tileOffset = 0
-			return m, nil
-		case "t":
-			name := CycleTheme()
-			return m, m.persistThemeCmd(name)
-		case "w":
-			return m.cycleTimeWindow()
-		case "v":
-			if m.screen == screenDashboard {
-				m.setDashboardView(m.nextDashboardView(1))
-				return m, m.persistDashboardViewCmd()
-			}
-		case "V":
-			if m.screen == screenDashboard {
-				m.setDashboardView(m.nextDashboardView(-1))
-				return m, m.persistDashboardViewCmd()
-			}
-		}
-	}
-
-	switch m.screen {
-	case screenAnalytics:
-		return m.handleAnalyticsKey(msg)
-	default:
-		return m.handleDashboardTilesKey(msg)
-	}
-}
-
-func (m Model) handleDashboardTilesKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	if m.filter.active {
-		return m.handleFilterKey(msg)
-	}
-	if m.mode == modeDetail {
-		return m.handleDetailKey(msg)
-	}
-	if m.activeDashboardView() == dashboardViewSplit {
-		return m.handleListKey(msg)
-	}
-	return m.handleTilesKey(msg)
-}
-
-func (m Model) handleAnalyticsKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	if m.analyticsFilter.active {
-		return m.handleAnalyticsFilterKey(msg)
-	}
-
-	switch msg.String() {
-	case "q", "ctrl+c":
-		return m, tea.Quit
-	case "s":
-		m.analyticsSortBy = (m.analyticsSortBy + 1) % analyticsSortCount
-	case "/":
-		m.analyticsFilter.active = true
-		m.analyticsFilter.text = ""
-	case "esc":
-		if m.analyticsFilter.text != "" {
-			m.analyticsFilter.text = ""
-		}
-	case "r":
-		m = m.requestRefresh()
-	}
-	return m, nil
-}
-
-func (m Model) handleAnalyticsFilterKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "enter":
-		m.analyticsFilter.active = false
-	case "esc":
-		m.analyticsFilter.active = false
-		m.analyticsFilter.text = ""
-	case "backspace":
-		if len(m.analyticsFilter.text) > 0 {
-			m.analyticsFilter.text = m.analyticsFilter.text[:len(m.analyticsFilter.text)-1]
-		}
-	default:
-		if len(msg.String()) == 1 {
-			m.analyticsFilter.text += msg.String()
-		}
-	}
-	return m, nil
-}
-
-func (m Model) availableScreens() []screenTab {
-	if !m.experimentalAnalytics {
-		return []screenTab{screenDashboard}
-	}
-	return []screenTab{screenDashboard, screenAnalytics}
-}
-
-func (m Model) nextScreen(step int) screenTab {
-	screens := m.availableScreens()
-	if len(screens) == 0 {
-		return screenDashboard
-	}
-
-	idx := 0
-	for i, screen := range screens {
-		if screen == m.screen {
-			idx = i
-			break
-		}
-	}
-
-	next := (idx + step) % len(screens)
-	if next < 0 {
-		next += len(screens)
-	}
-	return screens[next]
-}
-
-func (m Model) handleListKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	ids := m.filteredIDs()
-	pageStep := m.listPageStep()
-	switch msg.String() {
-	case "q", "ctrl+c":
-		return m, tea.Quit
-	case "up", "k":
-		if m.cursor > 0 {
-			m.cursor--
-			m.detailOffset = 0
-			m.detailTab = 0
-			m.tileOffset = 0
-		}
-	case "down", "j":
-		if m.cursor < len(ids)-1 {
-			m.cursor++
-			m.detailOffset = 0
-			m.detailTab = 0
-			m.tileOffset = 0
-		}
-	case "pgdown", "ctrl+d":
-		if len(ids) > 0 {
-			m.cursor = clamp(m.cursor+pageStep, 0, len(ids)-1)
-		}
-	case "pgup", "ctrl+u":
-		if len(ids) > 0 {
-			m.cursor = clamp(m.cursor-pageStep, 0, len(ids)-1)
-		}
-	case "enter", "right", "l":
-		m.mode = modeDetail
-		m.detailOffset = 0
-	case "/":
-		m.filter.active = true
-		m.filter.text = ""
-	case "r":
-		m = m.requestRefresh()
-	}
-	return m, nil
-}
-
-func (m Model) handleDetailKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "q", "ctrl+c":
-		return m, tea.Quit
-	case "esc", "left", "h", "backspace":
-		m.mode = modeList
-	case "up", "k":
-		if m.detailOffset > 0 {
-			m.detailOffset--
-		}
-	case "down", "j":
-		m.detailOffset++ // capped during render
-	case "g":
-		m.detailOffset = 0
-	case "G":
-		m.detailOffset = 9999 // will be capped
-	case "[":
-		if m.detailTab > 0 {
-			m.detailTab--
-			m.detailOffset = 0
-		}
-	case "]":
-		m.detailTab++
-		m.detailOffset = 0
-	case "1", "2", "3", "4", "5", "6", "7", "8", "9":
-		idx := int(msg.String()[0] - '1') // "1" → 0, "2" → 1, ...
-		m.detailTab = idx
-		m.detailOffset = 0
-	}
-	return m, nil
-}
-
-func (m Model) handleFilterKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "enter":
-		m.filter.active = false
-		m.cursor = 0
-		m.tileOffset = 0
-	case "esc":
-		m.filter.text = ""
-		m.filter.active = false
-		m.cursor = 0
-		m.tileOffset = 0
-	case "backspace":
-		if len(m.filter.text) > 0 {
-			m.filter.text = m.filter.text[:len(m.filter.text)-1]
-		}
-	default:
-		if len(msg.String()) == 1 {
-			m.filter.text += msg.String()
-		}
-	}
-	return m, nil
-}
-
-func (m Model) handleTilesKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	ids := m.filteredIDs()
-	cols := m.tileCols()
-	scrollModeWidget := m.shouldUseWidgetScroll()
-	switch msg.String() {
-	case "q", "ctrl+c":
-		return m, tea.Quit
-	case "up", "k":
-		if m.cursor >= cols {
-			m.cursor -= cols
-			m.tileOffset = 0
-		}
-	case "down", "j":
-		if m.cursor+cols < len(ids) {
-			m.cursor += cols
-			m.tileOffset = 0
-		}
-	case "left", "h":
-		if m.cursor > 0 {
-			m.cursor--
-			m.tileOffset = 0
-		}
-	case "right", "l":
-		if m.cursor < len(ids)-1 {
-			m.cursor++
-			m.tileOffset = 0
-		}
-	case "pgdown", "ctrl+d":
-		if scrollModeWidget {
-			m.tileOffset += m.widgetScrollStep()
-		} else {
-			m.tileOffset += m.tileScrollStep()
-		}
-	case "pgup", "ctrl+u":
-		if scrollModeWidget {
-			m.tileOffset -= m.widgetScrollStep()
-		} else {
-			m.tileOffset -= m.tileScrollStep()
-		}
-		if m.tileOffset < 0 {
-			m.tileOffset = 0
-		}
-	case "ctrl+o":
-		if id := m.selectedTileID(ids); id != "" {
-			m.expandedModelMixTiles[id] = !m.expandedModelMixTiles[id]
-		}
-	case "home":
-		m.tileOffset = 0
-	case "end":
-		m.tileOffset = 9999 // capped during render
-	case "enter":
-		m.mode = modeDetail
-		m.detailOffset = 0
-	case "/":
-		m.filter.active = true
-		m.filter.text = ""
-	case "esc":
-		if m.filter.text != "" {
-			m.filter.text = ""
-			m.cursor = 0
-			m.tileOffset = 0
-		}
-	case "r":
-		m = m.requestRefresh()
-	}
-	return m, nil
-}
-
-func (m Model) cycleTimeWindow() (tea.Model, tea.Cmd) {
-	next := core.NextTimeWindow(m.timeWindow)
-	m = m.beginTimeWindowRefresh(next)
-	return m, m.persistTimeWindowCmd(string(next))
-}
-
-func (m Model) requestRefresh() Model {
-	m.refreshing = true
-	if m.onRefresh != nil {
-		m.onRefresh(m.timeWindow)
-	}
-	return m
-}
-
-func (m Model) beginTimeWindowRefresh(window core.TimeWindow) Model {
-	m.timeWindow = window
-	if m.onTimeWindowChange != nil {
-		m.onTimeWindowChange(window)
-	}
-	m.refreshing = true
-	if m.onRefresh != nil {
-		m.onRefresh(window)
-	}
-	return m
-}
 
 func (m Model) selectedTileID(ids []string) string {
 	if len(ids) == 0 {
@@ -1231,24 +336,6 @@ func (m Model) shouldUsePanelScroll() bool {
 	return m.tileCols() == 1
 }
 
-func snapshotsReady(snaps map[string]core.UsageSnapshot) bool {
-	if len(snaps) == 0 {
-		return false
-	}
-	for _, snap := range snaps {
-		if snap.Status != core.StatusUnknown {
-			return true
-		}
-		if len(snap.Metrics) > 0 ||
-			len(snap.Resets) > 0 ||
-			len(snap.DailySeries) > 0 ||
-			len(snap.ModelUsage) > 0 {
-			return true
-		}
-	}
-	return false
-}
-
 func (m Model) View() string {
 	if m.width < 30 || m.height < 8 {
 		return lipgloss.NewStyle().
@@ -1266,42 +353,6 @@ func (m Model) View() string {
 		return m.renderSettingsModalOverlay()
 	}
 	return view
-}
-
-func (m Model) installDaemonCmd() tea.Cmd {
-	fn := m.onInstallDaemon
-	return func() tea.Msg {
-		if fn == nil {
-			return daemonInstallResultMsg{err: fmt.Errorf("install callback not configured")}
-		}
-		return daemonInstallResultMsg{err: fn()}
-	}
-}
-
-func (m Model) renderDashboard() string {
-	w, h := m.width, m.height
-
-	header := m.renderHeader(w)
-	headerH := strings.Count(header, "\n") + 1
-
-	footer := m.renderFooter(w)
-	footerH := strings.Count(footer, "\n") + 1
-
-	contentH := h - headerH - footerH
-	if contentH < 3 {
-		contentH = 3
-	}
-
-	var content string
-
-	switch m.screen {
-	case screenAnalytics:
-		content = m.renderAnalyticsContent(w, contentH)
-	default:
-		content = m.renderDashboardContent(w, contentH)
-	}
-
-	return header + "\n" + content + "\n" + footer
 }
 
 func (m Model) renderDashboardContent(w, contentH int) string {
